fix: run only one embedded Hedera instance at a time
