+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/junit/hedera/embedded/AbstractEmbeddedHedera.java
@@ -88,6 +88,8 @@ public abstract class AbstractEmbeddedHedera implements EmbeddedHedera {
     protected final AtomicInteger nextNano = new AtomicInteger(0);
     protected final Metrics metrics;
     protected final ScheduledExecutorService executorService = Executors.newSingleThreadScheduledExecutor();
+    // Held so stop() can unregister it; otherwise the hook pins this instance for the life of the JVM
+    private final Thread shutdownHook = new Thread(executorService::shutdownNow);
 
     /**
      * A trigger for the Hedera platform to use when initializing the state.
@@ -147,7 +149,7 @@ public abstract class AbstractEmbeddedHedera implements EmbeddedHedera {
                 metricsConfig);
         state = new FakeState();
         rebuildHedera();
-        Runtime.getRuntime().addShutdownHook(new Thread(executorService::shutdownNow));
+        Runtime.getRuntime().addShutdownHook(shutdownHook);
     }
 
     @Override
@@ -188,6 +190,11 @@ public abstract class AbstractEmbeddedHedera implements EmbeddedHedera {
         fakePlatform().notifyListeners(FREEZE_COMPLETE_NOTIFICATION);
         hedera.newPlatformStatus(FREEZE_COMPLETE_NOTIFICATION.getNewStatus());
         executorService.shutdownNow();
+        try {
+            Runtime.getRuntime().removeShutdownHook(shutdownHook);
+        } catch (final IllegalStateException ignore) {
+            // The JVM is already shutting down
+        }
     }
 
     @Override
